Fix: use `@components/Head` in `_eror.js`
